oci: cas: implement Walk and refactor GC
GC implements a lot of really cool recursive descent functionality, but
it wasn't really exposed. Split out the core parts into Walk() -- which
is meant to mirror "path/filepath".Walk. GC will be reimplemented to use
the new functionality.

In addition, implement Reachable and Paths which return information
about the set of descriptors and digests which can be reached. This
change also includes replacing gc.mark with Reachable.
